Consolidated unit tests for container states into one file - TestContainerStateCounts
